Move from dist -> lib
Seems more usual
