+++ .gitignore
@@ -1,4 +1,4 @@
 node_modules
-dist
+lib
 .nyc_output
 coverage
