+++ package.json
@@ -2,8 +2,11 @@
     "name": "dopri",
     "version": "0.0.1",
     "description": "Dormand–Prince methods",
-    "main": "dist/index.js",
-    "types": "dist/index.d.ts",
+    "main": "lib/index.js",
+    "types": "lib/index.d.ts",
+    "files": [
+        "/lib"
+    ],
     "scripts": {
         "build": "tsc",
         "test": "nyc mocha",
